refactor: lights data and service simplified a lot
