+++ src/data/lights-spb.json
@@ -1,0 +1,74 @@
+[
+	16, 15, 9, 50,
+	16, 25, 9, 45,
+	16, 35, 9, 40,
+	16, 45, 9, 30,
+	17, 0, 9, 20,
+	17, 10, 9, 10,
+	17, 25, 9, 0,
+	17, 40, 8, 50,
+	17, 50, 8, 35,
+	18, 5, 8, 20,
+	18, 15, 8, 10,
+	18, 25, 8, 0,
+	18, 40, 7, 45,
+	18, 50, 7, 30,
+	19, 5, 7, 15,
+	19, 15, 7, 0,
+	19, 30, 6, 30,
+	19, 40, 6, 10,
+	20, 5, 5, 50,
+	20, 25, 5, 35,
+	20, 35, 5, 20,
+	20, 50, 5, 5,
+	21, 5, 4, 50,
+	21, 20, 4, 30,
+	21, 30, 4, 15,
+	21, 50, 4, 0,
+	22, 0, 3, 45,
+	22, 15, 3, 30,
+	22, 30, 3, 15,
+	22, 45, 3, 0,
+	23, 0, 2, 45,
+	23, 5, 2, 35,
+	23, 15, 2, 30,
+	23, 15, 2, 30,
+	23, 15, 2, 35,
+	23, 15, 2, 40,
+	23, 10, 2, 50,
+	23, 0, 2, 55,
+	22, 50, 3, 10,
+	22, 40, 3, 20,
+	22, 25, 3, 35,
+	22, 10, 3, 55,
+	21, 55, 4, 10,
+	21, 40, 4, 25,
+	21, 20, 4, 40,
+	21, 5, 4, 55,
+	20, 50, 5, 5,
+	20, 30, 5, 20,
+	20, 15, 5, 35,
+	20, 0, 5, 45,
+	19, 45, 6, 0,
+	19, 30, 6, 10,
+	19, 15, 6, 25,
+	19, 0, 6, 40,
+	18, 30, 7, 0,
+	18, 10, 7, 20,
+	17, 50, 7, 35,
+	17, 35, 7, 45,
+	17, 25, 8, 0,
+	17, 10, 8, 10,
+	16, 55, 8, 25,
+	16, 45, 8, 40,
+	16, 35, 8, 50,
+	16, 20, 9, 0,
+	16, 15, 9, 10,
+	16, 10, 9, 20,
+	16, 0, 9, 30,
+	16, 0, 9, 40,
+	16, 0, 9, 40,
+	16, 0, 9, 50,
+	16, 0, 9, 50,
+	16, 1, 9, 50
+]
