--- src/lib/pages/illumination/illumination.page.svelte
@@ -1,7 +1,0 @@
-<script lang="ts">
-	import Today from "./today.svelte";
-	import IlluminationByDate from "./info-by-date.svelte";
-</script>
-
-<Today />
-<IlluminationByDate />
