+++ src/lib/pages/lights/lights.page.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+	import LightsToday from "./lights-today.svelte";
+	import LightsByDate from "./lights-schedule.svelte";
+</script>
+
+<LightsToday />
+<LightsByDate />
