+++ src/lib/services/lights/index.ts
@@ -1,0 +1,4 @@
+export {
+	getLightsScheduleByDateSPb as getLightsScheduleByDate,
+	getLightsStateByDateSpb as getLightsStateByDate
+} from "./illumination-spb";
