--- src/lib/types.ts
@@ -1,7 +1,0 @@
-export type EventName =
-	| "lights:start"
-	| "lights:end"
-	| "moonrise"
-	| "moonset"
-	| "sunrise"
-	| "sunset";
